requirements.txt
discord.py
py-trello
requests_oauthlib

